Change default native type to DPI_ORACLE_TYPE_INT64 if the column metadata indicates that the values are able to fit inside a 64-bit integer.

+++ samples/TestBLOB.c
@@ -101,8 +101,8 @@ int main(int argc, char **argv)
             return ShowError();
         if (dpiLob_getSize(blobColValue->value.asLOB, &blobSize) < 0)
             return ShowError();
-        printf("Row: IntCol = %g, BlobCol = BLOB(%" PRIu64 ")\n",
-                intColValue->value.asDouble, blobSize);
+        printf("Row: IntCol = %" PRId64 ", BlobCol = BLOB(%" PRIu64 ")\n",
+                intColValue->value.asInt64, blobSize);
     }
 
     // display description of each variable
